fix GDAL grid loading when the file has a non-positive scale element
GDAL files can have, and commonly have when produced from GIS systems,
scale factors that are negative. This is used to orient the bitmap north-
south instead of south-north. However, it also means that the
geotransform might not be affine.

Handle this case. For the time being, this is done by completely rejecting
any grid that would require some rotation applied.