core: do not reset an item ID when detached
If needs be, one can reset it before re-attaching

  setUniqueId(Environment::ITEM_NOT_ATTACHED)

However, the two most common detach workflows are:
 * delete the environment
 * create a smaller environment from a bigger one

in the second case, the reinitialization of the IDs forbids using global
names when doing environment sharing. In the first case, it does not
matter either way. So, change the behaviour, and make sure that automatic
ID generation cannot generate collisions.