Environment: add method for pulling binary events
One possible use case for the environment is insight a custom library.
To simplify the synchronization of these environments with a viewer
a method was added to pull all events as binary events. This method is
more or less the opposite to applyEvents

environment->pullEvents(events)
viewer.environment->applyEvents(events,all==true)