grid: introduce BandedGrid class in class hierarchy
this is meant to be used for template independent code.
Implemented is the createBand method, which allows to create named
bands without knowing the derived type of the object. This was
necessary since we don't want to always generate possible bands, but
only those that are actually used. To make this available to the tools
for creating grid based classes, this polymorphic approach is necessary.